feat: split control actions logic to 'fragment'
