+++ lib/Controls/ActionRecorder.js
@@ -405,14 +405,14 @@ export default class ActionRecorder extends CoreBase {
 			if (!control) throw new Error(`Unknown control: ${controlId}`)
 
 			if (mode === 'append') {
-				if (typeof control.actionAppend === 'function') {
-					if (!control.actionAppend(setId, this.#currentSession.actions)) throw new Error('Unknown set')
+				if (control.actions && typeof control.actions.actionAppend === 'function') {
+					if (!control.actions.actionAppend(setId, this.#currentSession.actions)) throw new Error('Unknown set')
 				} else {
 					throw new Error('Not supported by control')
 				}
 			} else {
-				if (typeof control.actionReplaceAll === 'function') {
-					if (!control.actionReplaceAll(setId, this.#currentSession.actions)) throw new Error('Unknown set')
+				if (control.actions && typeof control.actions.actionReplaceAll === 'function') {
+					if (!control.actions.actionReplaceAll(setId, this.#currentSession.actions)) throw new Error('Unknown set')
 				} else {
 					throw new Error('Not supported by control')
 				}
