+++ lib/Controls/ControlTypes/Button/Base.js
@@ -2,6 +2,7 @@ import ControlBase from '../../ControlBase.js'
 import { ParseControlId, rgb, clamp, GetButtonBitmapSize } from '../../../Resources/Util.js'
 import { cloneDeep, isEqual } from 'lodash-es'
 import { nanoid } from 'nanoid'
+import FragmentActions from '../../Fragments/FragmentActions.js'
 
 /**
  * Abstract class for a editable button control.
@@ -53,10 +54,10 @@ export default class ButtonControlBase extends ControlBase {
 	}
 
 	/**
-	 * The action-sets on this button
+	 * The actions fragment
 	 * @access public
 	 */
-	action_sets = {}
+	actions
 
 	/**
 	 * The current status of this button
@@ -99,326 +100,15 @@ export default class ButtonControlBase extends ControlBase {
 	 */
 	pushed = false
 
-	/**
-	 * Add an action to this control
-	 * @param {string} setId
-	 * @param {object} actionItem
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	actionAdd(setId, actionItem) {
-		if (this.action_sets[setId] === undefined) {
-			// cant implicitly create a set
-			this.logger.silly(`Missing set ${this.controlId}:${setId}`)
-			return false
-		}
+	constructor(registry, controlId, logSource, debugNamespace) {
+		super(registry, controlId, logSource, debugNamespace)
 
-		this.action_sets[setId].push(actionItem)
-
-		this.#actionSubscribe(actionItem)
-
-		this.commitChange(false)
-
-		this.checkButtonStatus()
-	}
-
-	/**
-	 * Duplicate an action on this control
-	 * @param {string} setId
-	 * @param {string} id
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	actionDuplicate(setId, id) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			const index = action_set.findIndex((act) => act.id === id)
-			if (index !== -1) {
-				const actionItem = cloneDeep(action_set[index])
-				actionItem.id = nanoid()
-
-				this.action_sets[setId].splice(index + 1, 0, actionItem)
-
-				this.#actionSubscribe(actionItem)
-
-				this.commitChange(false)
-
-				this.checkButtonStatus()
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Enable or disable an action
-	 * @param {string} setId
-	 * @param {string} id
-	 * @param {boolean} enabled
-	 * @access public
-	 */
-	actionEnabled(setId, id, enabled) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			for (const action of action_set) {
-				if (action && action.id === id) {
-					if (!action.options) action.options = {}
-
-					action.disabled = !enabled
-
-					// Inform relevant module
-					if (!action.disabled) {
-						this.#actionSubscribe(action)
-					} else {
-						this.cleanupAction(action)
-					}
-
-					this.commitChange(false)
-
-					return true
-				}
-			}
-		}
-	}
-
-	/**
-	 * Learn the options for an action, by asking the instance for the current values
-	 * @param {string} setId the id of the action set
-	 * @param {string} id the id of the action
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	async actionLearn(setId, id) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			const action = action_set.find((act) => act.id === id)
-			if (action) {
-				const instance = this.instance.moduleHost.getChild(action.instance)
-				if (instance) {
-					const newOptions = await instance.actionLearnValues(action)
-					if (newOptions) {
-						const newAction = {
-							...action,
-							options: newOptions,
-						}
-
-						// It may not still exist, so do a replace through the usual flow
-						return this.actionReplace(newAction)
-					}
-				}
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Remove an action from this control
-	 * @param {string} setId the id of the action set
-	 * @param {string} id the id of the action
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	actionRemove(setId, id) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			const index = action_set.findIndex((act) => act.id === id)
-			if (index !== -1) {
-				const action = action_set[index]
-				action_set.splice(index, 1)
-
-				this.cleanupAction(action)
-
-				this.commitChange(false)
-
-				this.checkButtonStatus()
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Reorder an action in the list or move between sets
-	 * @param {string} dragSetId the action_set id to remove from
-	 * @param {number} dragIndex the index of the action to move
-	 * @param {string} dropSetId the target action_set of the action
-	 * @param {number} dropIndex the target index of the action
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	actionReorder(dragSetId, dragIndex, dropSetId, dropIndex) {
-		const fromSet = this.action_sets[dragSetId]
-		const toSet = this.action_sets[dropSetId]
-		if (fromSet && toSet) {
-			dragIndex = clamp(dragIndex, 0, fromSet.length)
-			dropIndex = clamp(dropIndex, 0, toSet.length)
-
-			toSet.splice(dropIndex, 0, ...fromSet.splice(dragIndex, 1))
-		}
-
-		this.commitChange()
-	}
-
-	/**
-	 * Replace a action with an updated version
-	 * @param {object} newProps
-	 * @access public
-	 */
-	actionReplace(newProps) {
-		for (const action_set of Object.values(this.action_sets)) {
-			for (const action of action_set) {
-				// Replace the new feedback in place
-				if (action.id === newProps.id) {
-					action.action = newProps.actionId
-					action.options = newProps.options
-
-					delete action.upgradeIndex
-
-					// Inform relevant module
-					this.#actionSubscribe(action)
-
-					this.commitChange(false)
-
-					return true
-				}
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Append some actions to this button
-	 * @param {string} setId the action_set id to update
-	 * @param {Array} newActions actions to append
-	 * @access public
-	 */
-	actionAppend(setId, newActions) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			// Add new actions
-			for (const action of newActions) {
-				this.action_sets[setId].push(action)
-
-				this.#actionSubscribe(action)
-			}
-
-			this.commitChange(false)
-
-			return true
-		}
-
-		return false
-	}
-
-	/**
-	 * Replace all the actions in a set
-	 * @param {string} setId the action_set id to update
-	 * @param {Array} newActions actions to populate
-	 * @access public
-	 */
-	actionReplaceAll(setId, newActions) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			// Remove the old actions
-			for (const action of action_set) {
-				this.cleanupAction(action)
-			}
-			this.action_sets[setId] = []
-
-			// Add new actions
-			for (const action of newActions) {
-				this.action_sets[setId].push(action)
-
-				this.#actionSubscribe(action)
-			}
-
-			this.commitChange(false)
-
-			return true
-		}
-
-		return false
-	}
-
-	/**
-	 * Set the delay of an action
-	 * @param {string} setId the action_set id
-	 * @param {string} id the action id
-	 * @param {number} delay the desired delay
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	actionSetDelay(setId, id, delay) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			for (const action of action_set) {
-				if (action && action.id === id) {
-					delay = Number(delay)
-					if (isNaN(delay)) delay = 0
-
-					action.delay = delay
-
-					this.commitChange(false)
-
-					return true
-				}
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Set an opton of an action
-	 * @param {string} setId the action_set id
-	 * @param {string} id the action id
-	 * @param {string} key the desired option to set
-	 * @param {any} value the new value of the option
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	actionSetOption(setId, id, key, value) {
-		const action_set = this.action_sets[setId]
-		if (action_set) {
-			for (const action of action_set) {
-				if (action && action.id === id) {
-					if (!action.options) action.options = {}
-
-					action.options[key] = value
-
-					// Inform relevant module
-					this.#actionSubscribe(action)
-
-					this.commitChange(false)
-
-					return true
-				}
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Inform the instance of an updated action
-	 * @param {object} action the action which changed
-	 * @access private
-	 */
-	#actionSubscribe(action) {
-		if (!action.disabled) {
-			const instance = this.instance.moduleHost.getChild(action.instance)
-			if (instance) {
-				instance.actionUpdate(action, this.controlId).catch((e) => {
-					this.logger.silly(`action_update to connection failed: ${e.message}`)
-				})
-			}
-		}
+		this.actions = new FragmentActions(
+			registry,
+			controlId,
+			this.commitChange.bind(this),
+			this.checkButtonStatus.bind(this)
+		)
 	}
 
 	/**
@@ -430,7 +120,7 @@ export default class ButtonControlBase extends ControlBase {
 	checkButtonStatus(redraw = true) {
 		// Find all the instances referenced by the bank
 		const instance_ids = new Set()
-		for (const actions in Object.values(this.action_sets)) {
+		for (const actions in Object.values(this.actions.action_sets)) {
 			for (const action of actions) {
 				instance_ids.add(action.instance)
 			}
@@ -466,21 +156,6 @@ export default class ButtonControlBase extends ControlBase {
 	}
 
 	/**
-	 * Inform the instance of a removed action
-	 * @param {object} action the action being removed
-	 * @access protected
-	 */
-	cleanupAction(action) {
-		// Inform relevant module
-		const instance = this.instance.moduleHost.getChild(action.instance)
-		if (instance) {
-			instance.actionDelete(action).catch((e) => {
-				this.logger.silly(`action_delete to connection failed: ${e.message}`)
-			})
-		}
-	}
-
-	/**
 	 * Inform the instance of a removed feedback
 	 * @param {object} fedeback the feedback being removed
 	 * @access private
@@ -510,12 +185,7 @@ export default class ButtonControlBase extends ControlBase {
 			this.#cleanupFeedback(feedback)
 		}
 
-		// Inform modules of action cleanup
-		for (const action_set of Object.values(this.action_sets)) {
-			for (const action of action_set) {
-				this.cleanupAction(action)
-			}
-		}
+		this.actions.destroy()
 
 		super.destroy()
 	}
@@ -596,9 +266,9 @@ export default class ButtonControlBase extends ControlBase {
 	async feedbackLearn(id) {
 		const feedback = this.feedbacks.find((fb) => fb.id === id)
 		if (feedback) {
-			const instance = this.instance.moduleHost.getChild(action.instance)
+			const instance = this.instance.moduleHost.getChild(feedback.instance_id)
 			if (instance) {
-				const newOptions = await instance.feedbackLearnValues(action)
+				const newOptions = await instance.feedbackLearnValues(feedback)
 				if (newOptions) {
 					const newFeedback = {
 						...feedback,
@@ -811,36 +481,23 @@ export default class ButtonControlBase extends ControlBase {
 	 * @access public
 	 */
 	forgetInstance(instanceId) {
-		let changed = false
+		let changedFeedbacks = false
 
 		// Cleanup any feedbacks
 		const newFeedbacks = []
 		for (const feedback of this.feedbacks) {
 			if (feedback.instance_id === instanceId) {
 				this.#cleanupFeedback(feedback)
-				changed = true
+				changedFeedbacks = true
 			} else {
 				newFeedbacks.push(feedback)
 			}
 		}
 		this.feedbacks = newFeedbacks
 
-		// Cleanup any actions
-		for (const [setId, action_set] of Object.entries(this.action_sets)) {
-			const newActions = []
-			for (const action of action_set) {
-				if (action.instance === instanceId) {
-					this.cleanupAction(action)
-					changed = true
-				} else {
-					newActions.push(action)
-				}
-			}
+		const changedActions = this.actions.forgetInstance(instanceId)
 
-			this.action_sets[setId] = newActions
-		}
-
-		if (changed) {
+		if (changedActions || changedFeedbacks) {
 			this.checkButtonStatus(false)
 
 			this.commitChange()
@@ -961,24 +618,8 @@ export default class ButtonControlBase extends ControlBase {
 				}
 			}
 		}
-		for (const action_set of Object.values(this.action_sets)) {
-			for (let i = 0; i < action_set.length; i++) {
-				const action = action_set[i]
-				action.id = nanoid()
 
-				if (action.instance === 'internal') {
-					const newAction = this.internalModule.actionUpgrade(action, this.controlId)
-					if (newAction) {
-						action_set[i] = newAction
-					}
-				} else {
-					const instance = this.instance.moduleHost.getChild(action.instance)
-					if (instance) {
-						ps.push(instance.actionUpdate(action, this.controlId))
-					}
-				}
-			}
-		}
+		ps.push(this.actions.postProcessImport())
 
 		Promise.all(ps).catch((e) => {
 			this.logger.silly(`posProcessImport for ${this.controlId} failed: ${e.message}`)
@@ -1127,14 +768,7 @@ export default class ButtonControlBase extends ControlBase {
 		this.feedbacks = this.feedbacks.filter((feedback) => !!feedback && knownInstanceIds.has(feedback.instance_id))
 		changed = changed || this.feedbacks.length !== feedbackLength
 
-		// Clean out actions
-		for (const setId of Object.keys(this.action_sets)) {
-			const lengthBefore = this.action_sets[setId].length
-			this.action_sets[setId] = this.action_sets[setId].filter(
-				(action) => !!action && knownInstanceIds.has(action.instance)
-			)
-			changed = changed || this.action_sets[setId].length !== lengthBefore
-		}
+		changed = changed || this.actions.verifyInstanceIds(knownInstanceIds)
 
 		if (changed) {
 			this.commitChange()
