+++ lib/Controls/ControlTypes/Button/Press.js
@@ -38,7 +38,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 		this.style = cloneDeep(ButtonControlBase.DefaultStyle)
 		this.options = cloneDeep(ButtonControlBase.DefaultOptions)
 		this.feedbacks = []
-		this.action_sets = {
+		this.actions.action_sets = {
 			down: [],
 			up: [],
 		}
@@ -55,7 +55,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 			this.style = storage.style || this.style
 			this.options = storage.options || this.options
 			this.feedbacks = storage.feedbacks || this.feedbacks
-			this.action_sets = storage.action_sets || this.action_sets
+			this.actions.action_sets = storage.action_sets || this.actions.action_sets
 
 			if (isImport) this.postProcessImport()
 		}
@@ -74,7 +74,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 		// if (changed) { // TODO - this would be good but 'breaks' old behaviour
 		const action_set_id = pressed ? 'down' : 'up'
 
-		const actions = this.action_sets[action_set_id]
+		const actions = this.actions.action_sets[action_set_id]
 		if (actions) {
 			this.logger.silly('found actions')
 
@@ -99,7 +99,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 			style: this.style,
 			options: this.options,
 			feedbacks: this.feedbacks,
-			action_sets: this.action_sets,
+			action_sets: this.actions.action_sets,
 		}
 		return clone ? cloneDeep(obj) : obj
 	}
