+++ lib/Controls/ControlTypes/Button/Step.js
@@ -47,7 +47,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 			stepAutoProgress: true,
 		}
 		this.feedbacks = []
-		this.action_sets = {
+		this.actions.action_sets = {
 			0: [],
 		}
 		this.#current_step_id = '0'
@@ -64,9 +64,9 @@ export default class ControlButtonStep extends ButtonControlBase {
 			this.style = storage.style || this.style
 			this.options = storage.options || this.options
 			this.feedbacks = storage.feedbacks || this.feedbacks
-			this.action_sets = storage.action_sets || this.action_sets
+			this.actions.action_sets = storage.action_sets || this.actions.action_sets
 
-			this.#current_step_id = Object.keys(this.action_sets)[0]
+			this.#current_step_id = Object.keys(this.actions.action_sets)[0]
 
 			if (isImport) this.postProcessImport()
 		}
@@ -80,17 +80,17 @@ export default class ControlButtonStep extends ButtonControlBase {
 	actionSetAdd() {
 		let redraw = false
 
-		const existingKeys = Object.keys(this.action_sets)
+		const existingKeys = Object.keys(this.actions.action_sets)
 			.map((k) => Number(k))
 			.filter((k) => !isNaN(k))
 		if (existingKeys.length === 0) {
 			// add the default '0' set
-			this.action_sets['0'] = []
+			this.actions.action_sets['0'] = []
 			redraw = true
 		} else {
 			// add one after the last
 			const max = Math.max(...existingKeys)
-			this.action_sets[`${max + 1}`] = []
+			this.actions.action_sets[`${max + 1}`] = []
 		}
 
 		this.commitChange(redraw)
@@ -106,7 +106,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 	 */
 	actionSetAdvanceDelta(amount) {
 		if (amount && typeof amount === 'number') {
-			const all_steps = Object.keys(this.action_sets)
+			const all_steps = Object.keys(this.actions.action_sets)
 			if (all_steps.length > 0) {
 				const current = all_steps.indexOf(this.#current_step_id)
 
@@ -130,7 +130,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 	 */
 	actionSetMakeCurrent(index) {
 		if (typeof index === 'number') {
-			const setId = Object.keys(this.action_sets)[index - 1]
+			const setId = Object.keys(this.actions.action_sets)[index - 1]
 			if (setId !== undefined) {
 				return this.actionSetSelectNext(setId)
 			}
@@ -146,19 +146,19 @@ export default class ControlButtonStep extends ButtonControlBase {
 	 * @access public
 	 */
 	actionSetRemove(setId) {
-		const oldKeys = Object.keys(this.action_sets)
+		const oldKeys = Object.keys(this.actions.action_sets)
 
 		if (oldKeys.length > 1) {
 			let redraw = false
 
-			if (this.action_sets[setId]) {
+			if (this.actions.action_sets[setId]) {
 				// Inform modules of the change
-				for (const action of this.action_sets[setId]) {
-					this.cleanupAction(action)
+				for (const action of this.actions.action_sets[setId]) {
+					this.actions.cleanupAction(action)
 				}
 
 				// Assume it exists
-				delete this.action_sets[setId]
+				delete this.actions.action_sets[setId]
 
 				// Update the next step
 				const oldIndex = oldKeys.indexOf(setId)
@@ -194,7 +194,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 	 * @access public
 	 */
 	actionSetSelectNext(setId) {
-		if (this.action_sets[setId]) {
+		if (this.actions.action_sets[setId]) {
 			this.#current_step_id = setId
 
 			this.sendRuntimePropsChange()
@@ -215,10 +215,10 @@ export default class ControlButtonStep extends ButtonControlBase {
 	 * @access public
 	 */
 	actionSetSwap(setId1, setId2) {
-		if ((this.action_sets[setId1], this.action_sets[setId2])) {
-			const tmp = this.action_sets[setId1]
-			this.action_sets[setId1] = this.action_sets[setId2]
-			this.action_sets[setId2] = tmp
+		if ((this.actions.action_sets[setId1], this.actions.action_sets[setId2])) {
+			const tmp = this.actions.action_sets[setId1]
+			this.actions.action_sets[setId1] = this.actions.action_sets[setId2]
+			this.actions.action_sets[setId2] = tmp
 
 			this.commitChange(false)
 
@@ -234,7 +234,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 	 * @access public
 	 */
 	getActiveStepIndex() {
-		const out = Object.keys(this.action_sets).indexOf(this.#current_step_id)
+		const out = Object.keys(this.actions.action_sets).indexOf(this.#current_step_id)
 		return out !== -1 ? out : undefined
 	}
 
@@ -264,7 +264,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 		if (changed && pressed) {
 			// ignore the up, as we don't perform any actions
 			const this_step = this.#current_step_id
-			const steps = Object.keys(this.action_sets)
+			const steps = Object.keys(this.actions.action_sets)
 			if (steps.length > 0) {
 				steps.sort()
 
@@ -280,7 +280,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 					this.sendRuntimePropsChange()
 				}
 
-				const actions = this.action_sets[action_set_id]
+				const actions = this.actions.action_sets[action_set_id]
 				if (actions) {
 					this.logger.silly('found actions')
 
@@ -306,7 +306,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 			style: this.style,
 			options: this.options,
 			feedbacks: this.feedbacks,
-			action_sets: this.action_sets,
+			action_sets: this.actions.action_sets,
 		}
 		return clone ? cloneDeep(obj) : obj
 	}
