+++ lib/Controls/Controller.js
@@ -327,10 +327,10 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionAdd === 'function') {
+			if (control.actions && typeof control.actions.actionAdd === 'function') {
 				const actionItem = this.instance.definitions.createActionItem(instanceId, actionId)
 				if (actionItem) {
-					return control.actionAdd(setId, actionItem)
+					return control.actions.actionAdd(setId, actionItem)
 				} else {
 					return false
 				}
@@ -343,8 +343,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionLearn === 'function') {
-				return control.actionLearn(setId, id)
+			if (control.actions && typeof control.actions.actionLearn === 'function') {
+				return control.actions.actionLearn(setId, id)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
@@ -354,8 +354,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionEnabled === 'function') {
-				return control.actionEnabled(setId, id, enabled)
+			if (control.actions && typeof control.actions.actionEnabled === 'function') {
+				return control.actions.actionEnabled(setId, id, enabled)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
@@ -365,8 +365,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionRemove === 'function') {
-				return control.actionRemove(setId, id)
+			if (control.actions && typeof control.actions.actionRemove === 'function') {
+				return control.actions.actionRemove(setId, id)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
@@ -376,8 +376,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionDuplicate === 'function') {
-				return control.actionDuplicate(setId, id)
+			if (control.actions && typeof control.actions.actionDuplicate === 'function') {
+				return control.actions.actionDuplicate(setId, id)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
@@ -387,8 +387,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionSetDelay === 'function') {
-				return control.actionSetDelay(setId, id, delay)
+			if (control.actions && typeof control.actions.actionSetDelay === 'function') {
+				return control.actions.actionSetDelay(setId, id, delay)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
@@ -398,8 +398,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionSetOption === 'function') {
-				return control.actionSetOption(setId, id, key, value)
+			if (control.actions && typeof control.actions.actionSetOption === 'function') {
+				return control.actions.actionSetOption(setId, id, key, value)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
@@ -408,8 +408,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.actionReorder === 'function') {
-				return control.actionReorder(dragSetId, dragIndex, dropSetId, dropIndex)
+			if (control.actions && typeof control.actions.actionReorder === 'function') {
+				return control.actions.actionReorder(dragSetId, dragIndex, dropSetId, dropIndex)
 			} else {
 				throw new Error(`Control "${controlId}" does not support actions`)
 			}
