+++ lib/Data/ImportExport.js
@@ -127,7 +127,7 @@ class DataImportExport extends CoreBase {
 						instance_ids.add(feedback.instance_id)
 					}
 
-					for (const action_set of Object.values(control.action_sets || {})) {
+					for (const action_set of Object.values(control.actions.action_sets || {})) {
 						for (const action of action_set) {
 							instance_ids.add(action.instance)
 						}
@@ -527,8 +527,8 @@ class DataImportExport extends CoreBase {
 			}
 			control.feedbacks = newFeedbacks
 		}
-		if (control.action_sets) {
-			for (const [setId, action_set] of Object.entries(control.action_sets)) {
+		if (control.actions && control.actions.action_sets) {
+			for (const [setId, action_set] of Object.entries(control.actions.action_sets)) {
 				const newActions = []
 				for (const action of action_set) {
 					const instanceInfo = instanceIdMap[action?.instance]
@@ -538,7 +538,7 @@ class DataImportExport extends CoreBase {
 						newActions.push(action)
 					}
 				}
-				control.action_sets[setId] = newActions
+				control.actions.action_sets[setId] = newActions
 			}
 		}
