+++ lib/Instance/Wrapper.js
@@ -181,8 +181,8 @@ class SocketEventsHandler {
 
 		const allControls = this.registry.controls.getAllControls()
 		for (const [controlId, control] of Object.entries(allControls)) {
-			if (control.action_sets) {
-				for (const set of Object.values(control.action_sets)) {
+			if (control.actions && control.actions.action_sets) {
+				for (const set of Object.values(control.actions.action_sets)) {
 					for (const action of set) {
 						const parsed = ParseControlId(controlId)
 						if (action.instance == this.connectionId) {
@@ -638,7 +638,7 @@ class SocketEventsHandler {
 				const parsedControl = ParseControlId(action.controlId)
 				if (parsedControl?.type === 'bank') {
 					const control = this.registry.controls.getControl(action.controlId)
-					const found = control?.actionReplace?.(action) ?? false
+					const found = control?.actions?.actionReplace?.(action) ?? false
 					if (!found) {
 						this.logger.silly(`Failed to replace upgraded action: ${action.id} ${action.controlId}`)
 					}
