+++ lib/Internal/Controller.js
@@ -66,14 +66,14 @@ export default class InternalController extends CoreBase {
 			}
 
 			// Discover actions to process
-			if (control.action_sets) {
-				for (const action_set of Object.values(control.action_sets)) {
+			if (control.actions && control.actions.actionReplace) {
+				for (const action_set of Object.values(control.actions.action_sets)) {
 					for (const action of action_set) {
 						if (action.instance === 'internal') {
 							// Try and run an upgrade
 							const newAction = this.actionUpgrade(action)
 							if (newAction) {
-								control.actionReplace(newAction)
+								control.actions.actionReplace(newAction)
 							}
 						}
 					}
